% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADBR70}
\alias{ADBR70}
\title{DHS Births dataset}
\format{
A data frame with 2753 rows and 8 variables:
\describe{
  \item{v005}{Women individual sample weight}
  \item{v007}{Year of interview}
  \item{v008}{Date of interview (CMC)}
  \item{v021}{Primary sampling unit}
  \item{v022}{Sample strata for sampling error}
  \item{v025}{Type of residence urban/rural}
  \item{b3}{Date of birth (CMC)}
  \item{b7}{Age at death}
}
}
\source{
\url{https://dhsprogram.com/data/available-datasets.cfm}
}
\usage{
ADBR70
}
\description{
Example for a DHS data of births.
}
\keyword{datasets}
